Move most verifier code to a dedicated component in explorer
This in order to re-use it later for cardano transactions certifications.
