+++ mithril-explorer/src/components/CertificateModal/index.js
@@ -8,7 +8,7 @@ import RawJsonButton from "../RawJsonButton";
 import Stake from "../Stake";
 import ProtocolParameters from "../ProtocolParameters";
 import SignerTable from "../SignerTable";
-import VerifyCertificate from "../VerifyCertificate";
+import VerifyCertificateModal from "../VerifyCertificate/modal";
 
 export default function CertificateModal(props) {
   const [certificate, setCertificate] = useState({});
@@ -42,10 +42,6 @@ export default function CertificateModal(props) {
 
   function showPrevious() {
     props.onHashChange(certificate.previous_hash);
-  }
-
-  function verifyCertificate(certificate) {
-    setShowVerifyCertificateModal(true);
   }
 
   function close() {
@@ -169,7 +165,10 @@ export default function CertificateModal(props) {
         )}
       </Modal.Body>
       <Modal.Footer>
-        <Button size="sm" onClick={() => verifyCertificate(certificate)} className="text-break">
+        <Button
+          size="sm"
+          onClick={() => setShowVerifyCertificateModal(true)}
+          className="text-break">
           Verify certificate
         </Button>
         {certificate.genesis_signature !== "" ? (
@@ -183,7 +182,7 @@ export default function CertificateModal(props) {
         )}
         <RawJsonButton href={certificateEndpoint} size="sm" />
       </Modal.Footer>
-      <VerifyCertificate
+      <VerifyCertificateModal
         show={showVerifyCertificateModal}
         onClose={() => setShowVerifyCertificateModal(false)}
         certificateHash={certificate.hash}
